+++ lib/puppet/type/postconf_master.rb
@@ -29,9 +29,9 @@ Puppet::Type.newtype(:postconf_master) do
       case value
       when :undef, '-', nil
         '-'
-      when true, :true, 'true', :yes, 'yes', :y, 'y' # rubocop:disable Lint/BooleanSymbol
+      when true, :true, 'true', :yes, 'yes', :y, 'y'
         'y'
-      when false, :false, 'false', :no, 'no', :n, 'n' # rubocop:disable Lint/BooleanSymbol
+      when false, :false, 'false', :no, 'no', :n, 'n'
         'n'
       else
         raise ArgumentError, "Invalid value #{value.inspect}. Valid values are true, false, y, n, -."