+++ CHANGELOG.md
@@ -10,6 +10,9 @@ and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.
 - Convert to PDK module
 - Update OS version, Puppet version and dependencies
 
+### Fixed
+- Fix puppet-lint and rubocop offenses
+
 ## Release [1.0.2] - 2018-10-28
 This release fixes postfix reloading on config changes.